requirements.txt
Django==1.4
MySQL-python
#-e git://github.com/mozilla/django-piston.git#egg=django-piston
#south
#fabric
flup
django-test-utils
django-paging
django-tastypie
#simplejson==2.1.3
#django-email-extras
#Mock
#boto
